Update shipping_year when milestones are edited.
