+++ api/features_api.py
@@ -35,6 +35,7 @@ from internals.data_types import VerboseFeatureDict
 from internals import feature_helpers
 from internals import search
 from internals import search_fulltext
+from internals import stage_helpers
 from internals.user_models import AppUser
 import settings
 
@@ -180,9 +181,9 @@ class FeaturesAPI(basehandlers.EntitiesAPIHandler):
       self,
       stage_changes_list: list[dict[str, Any]],
       changed_fields: CHANGED_FIELDS_LIST_TYPE
-    ) -> bool:
+    ) -> list[Stage]:
     """Update stage fields with changes provided in the PATCH request."""
-    stages: list[Stage] = []
+    stages_to_store: list[Stage] = []
     for change_info in stage_changes_list:
       stage_was_updated = False
       # Check if valid ID is provided and fetch stage if it exists.
@@ -223,29 +224,32 @@ class FeaturesAPI(basehandlers.EntitiesAPIHandler):
       stage.milestones = milestones
 
       if stage_was_updated:
-        stages.append(stage)
+        stages_to_store.append(stage)
 
     # Save all of the updates made.
-    if stages:
-      ndb.put_multi(stages)
+    if stages_to_store:
+      ndb.put_multi(stages_to_store)
 
-    # Return a boolean representing if any changes were made to any stages.
-    return len(stages) > 0
+    # Return the list of modified stages.
+    return stages_to_store
 
   def _patch_update_special_fields(
       self,
       feature: FeatureEntry,
       feature_changes: dict[str, Any],
-      has_updated: bool
+      has_updated: bool,
+      updated_stages: list[Stage],
+      changed_fields: CHANGED_FIELDS_LIST_TYPE,
     ) -> None:
     """Handle any special FeatureEntry fields."""
     now = datetime.now()
-    # Handle accuracy notifications if this is an accuracy verification request.
+    # Set accurate_as_of if this is an accuracy verification request.
     if 'accurate_as_of' in feature_changes:
       feature.accurate_as_of = now
       feature.outstanding_notifications = 0
       has_updated = True
 
+    # Set enterprise first notification milestones.
     if is_update_first_notification_milestone(feature, feature_changes):
       feature.first_enterprise_notification_milestone = int(feature_changes['first_enterprise_notification_milestone'])
       has_updated = True
@@ -255,6 +259,30 @@ class FeaturesAPI(basehandlers.EntitiesAPIHandler):
     if should_remove_first_notice_milestone(feature, feature_changes):
       feature.first_enterprise_notification_milestone = None
 
+    # Set shipping_year based on milestones.
+    if updated_stages:
+      existing_shipping_stages = (
+          stage_helpers.get_all_shipping_stages_with_milestones(
+              feature_id=feature.key.integer_id()))
+      shipping_stage_dict = {
+          es.key.integer_id(): es
+          for es in existing_shipping_stages
+      }
+      shipping_stage_dict.update({
+          us.key.integer_id(): us
+          for us in updated_stages
+          if us.stage_type in STAGE_TYPES_SHIPPING.values()
+      })
+      earliest = stage_helpers.find_earliest_milestone(
+          list(shipping_stage_dict.values()))
+      if earliest:
+          year = stage_helpers.look_up_year(earliest)
+          if year != feature.shipping_year:
+            changed_fields.append(('shipping_year', feature.shipping_year, year))
+            feature.shipping_year = year
+            has_updated = True
+
+    # If changes were made, set the feature.updated timestamp.
     if has_updated:
       user_email = self.get_current_user().email()
       feature.updater_email = user_email
@@ -264,10 +292,11 @@ class FeaturesAPI(basehandlers.EntitiesAPIHandler):
       self,
       feature: FeatureEntry,
       feature_changes: dict[str, Any],
-      has_updated: bool,
+      updated_stages: list[Stage],
       changed_fields: CHANGED_FIELDS_LIST_TYPE
     ) -> None:
     """Update feature fields with changes provided in the PATCH request."""
+    has_updated = len(updated_stages) > 0
     for field, field_type in api_specs.FEATURE_FIELD_DATA_TYPES:
       if field not in feature_changes:
         continue
@@ -277,7 +306,8 @@ class FeaturesAPI(basehandlers.EntitiesAPIHandler):
       changed_fields.append((field, old_value, new_value))
       has_updated = True
 
-    self._patch_update_special_fields(feature, feature_changes, has_updated)
+    self._patch_update_special_fields(
+        feature, feature_changes, has_updated, updated_stages, changed_fields)
     feature.put()
 
   def do_patch(self, **kwargs):
@@ -299,9 +329,9 @@ class FeaturesAPI(basehandlers.EntitiesAPIHandler):
       return redirect_resp
 
     changed_fields: CHANGED_FIELDS_LIST_TYPE = []
-    has_updated = self._patch_update_stages(body['stages'], changed_fields)
+    updated_stages = self._patch_update_stages(body['stages'], changed_fields)
     self._patch_update_feature(
-      feature, body['feature_changes'], has_updated, changed_fields)
+      feature, body['feature_changes'], updated_stages, changed_fields)
 
     notifier_helpers.notify_subscribers_and_save_amendments(
         feature, changed_fields, notify=True)
